refactor: rename Model namespace as ApiModels
