+++ Doppler.HtmlEditorApi.Test/GetCampaignTest.cs
@@ -5,7 +5,7 @@ using System.Net.Http.Headers;
 using System.Text.Json;
 using System.Threading.Tasks;
 using Doppler.HtmlEditorApi.Infrastructure;
-using Doppler.HtmlEditorApi.Model;
+using Doppler.HtmlEditorApi.ApiModels;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.Extensions.DependencyInjection;
 using Moq;
