+++ Doppler.HtmlEditorApi/ApiModels/Content.cs
@@ -1,5 +1,5 @@
 using System.Text.Json;
 
-namespace Doppler.HtmlEditorApi.Model;
+namespace Doppler.HtmlEditorApi.ApiModels;
 
 public record Content(ContentType type, JsonElement? meta, string htmlContent);
