+++ Doppler.HtmlEditorApi/ApiModels/Template.cs
@@ -1,5 +1,5 @@
 using System.Text.Json;
 
-namespace Doppler.HtmlEditorApi.Model;
+namespace Doppler.HtmlEditorApi.ApiModels;
 
 public record Template(ContentType type, string name, JsonElement? meta, string htmlContent) : Content(type, meta, htmlContent);
