+++ Doppler.HtmlEditorApi/Controllers/TemplatesController.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using Doppler.HtmlEditorApi.DopplerSecurity;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Logging;
-using Doppler.HtmlEditorApi.Model;
+using Doppler.HtmlEditorApi.ApiModels;
 using Doppler.HtmlEditorApi.Infrastructure;
