+++ Doppler.HtmlEditorApi/Infrastructure/DummyRepository.cs
@@ -1,5 +1,4 @@
 using System;
-using Doppler.HtmlEditorApi.Model;
 using System.Threading.Tasks;
 using System.Text.Json;
 using System.Collections.Generic;
